setcustomfields
