+++ apps/meteor/app/livechat/server/api/v1/customField.ts
@@ -2,7 +2,7 @@ import { isLivechatCustomFieldsProps, isPOSTLivechatCustomFieldParams, isPOSTLiv
 
 import { API } from '../../../../api/server';
 import { getPaginationItems } from '../../../../api/server/helpers/getPaginationItems';
-import { Livechat } from '../../lib/LivechatTyped';
+import { setCustomFields } from '../../lib/custom-fields';
 import { findLivechatCustomFields, findCustomFieldById } from '../lib/customFields';
 import { findGuest } from '../lib/livechat';
 
@@ -18,7 +18,7 @@ API.v1.addRoute(
 				throw new Error('invalid-token');
 			}
 
-			if (!(await Livechat.setCustomFields({ token, key, value, overwrite }))) {
+			if (!(await setCustomFields({ token, key, value, overwrite }))) {
 				return API.v1.failure();
 			}
 
@@ -46,7 +46,7 @@ API.v1.addRoute(
 						overwrite: boolean;
 					}): Promise<{ Key: string; value: string; overwrite: boolean }> => {
 						const data = Object.assign({ token }, customField);
-						if (!(await Livechat.setCustomFields(data))) {
+						if (!(await setCustomFields(data))) {
 							throw new Error('error-setting-custom-field');
 						}
