--- apps/meteor/app/livechat/server/lib/validateRequiredCustomFields.ts
@@ -1,16 +1,0 @@
-import type { ILivechatCustomField } from '@rocket.chat/core-typings';
-
-export const validateRequiredCustomFields = (customFields: string[], livechatCustomFields: ILivechatCustomField[]) => {
-	const errors: string[] = [];
-	const requiredCustomFields = livechatCustomFields.filter((field) => field.required);
-
-	requiredCustomFields.forEach((field) => {
-		if (!customFields.find((f) => f === field._id)) {
-			errors.push(field._id);
-		}
-	});
-
-	if (errors.length > 0) {
-		throw new Error(`Missing required custom fields: ${errors.join(', ')}`);
-	}
-};
